+++ scripts/vk_validation_stats.py
@@ -61,6 +61,7 @@ generated_layer_source_files = [
 layer_source_files = [
 '../layers/buffer_validation.cpp',
 '../layers/core_validation.cpp',
+'../layers/core_validation_drawdispatch.cpp',
 '../layers/descriptor_sets.cpp',
 '../layers/parameter_validation_utils.cpp',
 '../layers/object_tracker_utils.cpp',
